==> All Features Random-Forest/Codes/01_Bioactivity_Data_Collection.ipynb <==
## **Installing libraries**

! pip install chembl_webresource_client

## **Importing libraries**

# Import necessary libraries
import pandas as pd
from chembl_webresource_client.new_client import new_client

## **Search for Target protein**\
# Prostaglandin E synthase

# Target search for coronavirus
target = new_client.target
target_query = target.search('CHEMBL5658')
targets = pd.DataFrame.from_dict(target_query)
targets

selected_target = targets.target_chembl_id[0]
selected_target

activity = new_client.activity
res = activity.filter(target_chembl_id=selected_target).filter(standard_type="IC50")

df = pd.DataFrame.from_dict(res)

df

df.standard_type.unique()

df.to_csv('PGES_bioactivity_data.csv', index=False)

## **Handling missing data**
If any compounds has missing value for the **standard_value** column then drop it

df2 = df[df.standard_value.notna()]
df2

## **Data pre-processing of the bioactivity data**

bioactivity_class = []
for i in df2.standard_value:
  if float(i) >= 10000:
    bioactivity_class.append("inactive")
  elif float(i) <= 1000:
    bioactivity_class.append("active")
  else:
    bioactivity_class.append("intermediate")

### **Iterate the *molecule_chembl_id* to a list**

mol_cid = []
for i in df2.molecule_chembl_id:
  mol_cid.append(i)

### **Iterate *canonical_smiles* to a list**

canonical_smiles = []
for i in df2.canonical_smiles:
  canonical_smiles.append(i)

### **Iterate *standard_value* to a list**

standard_value = []
for i in df2.standard_value:
  standard_value.append(i)

### **Combine the 4 lists into a dataframe**

data_tuples = list(zip(mol_cid, canonical_smiles, bioactivity_class, standard_value))
df3 = pd.DataFrame( data_tuples,  columns=['molecule_chembl_id', 'canonical_smiles', 'bioactivity_class', 'standard_value'])

df3

Saves dataframe to CSV file

df3.to_csv('01_PGES_bioactivity_preprocessed_data.csv', index=False)

---
